Day 2 solved

+++ src/main/java/de/v1per/adventofcode/days/Day1.java
@@ -6,16 +6,16 @@ import java.util.List;
 
 public class Day1 extends Day {
     public static void main(String[] args) {
-        List<Integer> intList = Utils.readFile("day1.txt", Integer.class);
+        List<Long> intList = Utils.readFile("day1.txt", Long.class);
         new Day1().play(intList);
     }
 
     @Override
     public void one() {
-        int prevNumber = 0;
+        long prevNumber = 0;
         //-1 because the first entry is a "false" increase
         int res = -1;
-        for (int number : (List<Integer>) list) {
+        for (long number : (List<Long>) list) {
             if (prevNumber < number) {
                 res++;
             }
@@ -26,12 +26,12 @@ public class Day1 extends Day {
 
     @Override
     public void two() {
-        int prevSum = 0;
+        long prevSum = 0;
         //-1 because the first entry is a "false" increase
         int res = -1;
         int i = 0;
         while (i + 2 < list.size()) {
-            int sum = (Integer) list.get(i) + (Integer) list.get(i + 1) + (Integer) list.get(i + 2);
+            long sum = (Long) list.get(i) + (Long) list.get(i + 1) + (Long) list.get(i + 2);
             if (prevSum < sum) {
                 res++;
             }
